+++ web/src/Services/ProjectService.ts
@@ -1,14 +1,9 @@
 import axios from 'axios';
+import { config, API_ENPOINT } from './config';
 
-var config = {
-    headers: {
-        'Access-Control-Allow-Origin': '*',
-        'Content-Type': 'application/json',
-    }
-};
 const ProjectService = {
     getProjects: async () => {
-        return (await axios.get('http://localhost:3000/projects', config)).data;
+        return (await axios.get(`${API_ENPOINT}/projects`, config)).data;
     }
 };
