+++ web/src/Services/index.ts
@@ -1,0 +1,6 @@
+import AuthService from "./AuthService";
+import NavService from "./NavigationService";
+import ProjectService from "./ProjectService";
+import PartService from "./PartService";
+
+export { AuthService, NavService, ProjectService, PartService }
